Fixed data type

+++ MoviesAndSeries.Server.API/Models/Information.cs
@@ -10,7 +10,7 @@
 
 		public string? Image { get; set; }
 
-		public ushort? Rating { get; set; }
+		public double? Rating { get; set; }
 
 		public uint Duration { get; set; }
 	}
